+++ R/tt_iterate.R
@@ -4,7 +4,7 @@
 #' targets to the target script. It allows the user to specify the input and output
 #' targets, as well as a custom user function to be applied.
 #'
-#' @param input_hpc A `tidytargets` object.
+#' @param tt_input A `tidytargets` object.
 #' @param target_output Character name of the output target. `NULL` uses an
 #'   auto-generated name.
 #' @param user_function A quoted function call to execute per iteration.
@@ -15,20 +15,20 @@
 #'   reference upstream targets by name.
 #'
 #' @export
-hpc_iterate <- function(
-    input_hpc,
+tt_iterate <- function(
+    tt_input,
     target_output = NULL,
     user_function = NULL,
     user_function_source_path = NULL,
     ...
 ) {
-  UseMethod("hpc_iterate")
+  UseMethod("tt_iterate")
 }
 
-#' @rdname hpc_iterate
+#' @rdname tt_iterate
 #' @export
-hpc_iterate.default <- function(
-    input_hpc,
+tt_iterate.default <- function(
+    tt_input,
     target_output = NULL,
     user_function = NULL,
     user_function_source_path = NULL,
@@ -37,13 +37,13 @@ hpc_iterate.default <- function(
   stop_if_not_tidytargets()
 }
 
-#' @rdname hpc_iterate
+#' @rdname tt_iterate
 #' @importFrom glue glue
 #' @importFrom magrittr %>%
 #' @importFrom purrr set_names
 #' @export
-hpc_iterate.tidytargets <- function(
-    input_hpc,
+tt_iterate.tidytargets <- function(
+    tt_input,
     target_output = NULL,
     user_function = NULL,
     user_function_source_path = NULL,
@@ -54,9 +54,9 @@ hpc_iterate.tidytargets <- function(
     check_for_name_value_conflicts(...)
     
     # Target script
-    target_script = glue("{input_hpc$initialisation$store}.R")
+    target_script = glue("{tt_input$initialisation$store}.R")
     
-    # Delete line with target in case the user execute the command, without calling hpc_initialise
+    # Delete line with target in case the user execute the command, without calling tt_initialise
     target_output |>  delete_lines_with_word(target_script)
     
     # Append source if any
@@ -65,24 +65,24 @@ hpc_iterate.tidytargets <- function(
     # please, because sometime we set up list target that do not depend on any other ones
     # if tiers is set to NULL, then the target will not acquire the _<tier> suffix
     # I HAVE TO MAKE THIS MORE ELEGANT, AND NOT RELY ON tiers ARGUMENT
-    if(input_hpc$initialisation$tier |> get_positions() |> length() < 2){
+    if(tt_input$initialisation$tier |> get_positions() |> length() < 2){
       iterate_value = "map"
       tiers_value = NULL
     }
       
     else if(
-      list(...) |> arguments_to_action(input_hpc, "tiered") |> length() == 0 &
-      list(...) |> arguments_to_action(input_hpc, "tier") |> length() == 0
+      list(...) |> arguments_to_action(tt_input, "tiered") |> length() == 0 &
+      list(...) |> arguments_to_action(tt_input, "tier") |> length() == 0
     ){
       iterate_value = "tier"
       tiers_value = NULL
     } else {
       iterate_value = "tiered"
-      tiers_value = input_hpc$initialisation$tier |> get_positions()
+      tiers_value = tt_input$initialisation$tier |> get_positions()
     }
 
     tar_append(
-      fx = hpc_factory |> quote(),
+      fx = tt_factory |> quote(),
       tiers = tiers_value ,
       target_output = target_output,
       script = target_script,
@@ -90,15 +90,15 @@ hpc_iterate.tidytargets <- function(
       
       # I HAVE TO IMPROVE the fact that I have to convert to character 
       # because arguments_to_action is also used in expand_tiered_arguments, which needs a named vector
-      arguments_to_tier = list(...) |> arguments_to_action(input_hpc, "tier") |> as.character()  , # This "tier" value is decided for each new target below. Usually just at the beginning of the piepline
-      arguments_already_tiered = list(...) |> arguments_to_action(input_hpc, "tiered") |> as.character(), # This "tiered" value is decided for each new target below. Ususally every other list targets.
-      other_arguments_to_map = list(...) |> arguments_to_action(input_hpc, c("tiered", "map")) |> as.character(), # This "tiered" value is decided for each new target below. Ususally every other list targets.
+      arguments_to_tier = list(...) |> arguments_to_action(tt_input, "tier") |> as.character()  , # This "tier" value is decided for each new target below. Usually just at the beginning of the piepline
+      arguments_already_tiered = list(...) |> arguments_to_action(tt_input, "tiered") |> as.character(), # This "tiered" value is decided for each new target below. Ususally every other list targets.
+      other_arguments_to_map = list(...) |> arguments_to_action(tt_input, c("tiered", "map")) |> as.character(), # This "tiered" value is decided for each new target below. Ususally every other list targets.
       ...
     )
   
       
     # Add pipeline step
-    input_hpc |>
+    tt_input |>
       c(
         as.list(environment())[-1] |> 
           c(list(iterate = iterate_value)) |>